# (*Stream).WriteMore: do not Flush